Rename database files
Use RecycleView for List and small refactor
